+++ cs2tracker/app/__init__.py
@@ -1,0 +1,3 @@
+from cs2tracker.app.application import (  # noqa: F401 # pylint:disable=unused-import
+    Application,
+)
