+++ cs2tracker/scraper/__init__.py
@@ -1,0 +1,9 @@
+from cs2tracker.scraper.background_task import (  # noqa: F401 # pylint:disable=unused-import
+    BackgroundTask,
+)
+from cs2tracker.scraper.discord_notifier import (  # noqa: F401 # pylint:disable=unused-import
+    DiscordNotifier,
+)
+from cs2tracker.scraper.scraper import (  # noqa: F401 # pylint:disable=unused-import
+    Scraper,
+)
